requirements.txt
scapy==2.5.0
pycryptodome==3.19.0
cryptography==41.0.7
flask==3.0.0